move component installation to cluster add
