+++ pkg/oc/bootstrap/clusteradd/componentinstall/apply_list.go
