+++ pkg/oc/bootstrap/clusteradd/componentinstall/apply_template.go
