+++ pkg/oc/bootstrap/clusteradd/componentinstall/privileged.go
