+++ pkg/oc/bootstrap/clusteradd/components/registry/OWNERS
@@ -1,0 +1,4 @@
+reviewers:
+  - bparees
+approvers:
+  - bparees
