+++ pkg/oc/bootstrap/clusteradd/components/router/OWNERS
@@ -1,0 +1,6 @@
+reviewers:
+  - danwinship
+  - dcbw
+approvers:
+  - danwinship
+  - dcbw
