+++ pkg/oc/bootstrap/clusteradd/components/service-catalog/OWNERS
@@ -1,0 +1,6 @@
+reviewers:
+  - jpeeler
+  - pmorie
+approvers:
+  - jpeeler
+  - pmorie
