+++ pkg/oc/bootstrap/clusteradd/components/template-service-broker/OWNERS
@@ -1,0 +1,6 @@
+reviewers:
+  - bparees
+  - jim-minter
+approvers:
+  - bparees
+  - jim-minter
