+++ pkg/oc/bootstrap/clusteradd/components/web-console/OWNERS
@@ -1,0 +1,4 @@
+reviewers:
+  - spadgett
+approvers:
+  - spadgett
