+++ pkg/oc/bootstrap/docker/openshift/templateservicebroker.go
@@ -8,7 +8,7 @@ import (
 	"github.com/golang/glog"
 
 	"github.com/openshift/origin/pkg/oc/bootstrap"
-	"github.com/openshift/origin/pkg/oc/bootstrap/clusterup/componentinstall"
+	"github.com/openshift/origin/pkg/oc/bootstrap/clusteradd/componentinstall"
 	"github.com/openshift/origin/pkg/oc/bootstrap/docker/errors"
 )
